+++ CHANGELOG.md
@@ -18,6 +18,7 @@ changes.
 - Holiday calculation methods in providers are now protected instead of private 
   to allow use in [custom providers](https://www.yasumi.dev/docs/cookbook/custom_provider/).
   [\#331](https://github.com/azuyalabs/yasumi/issues/331)
+- For Mexico, add the Day of Transmission of Federal Executive Power in election years. [\#361](https://github.com/azuyalabs/yasumi/pull/361) ([c960657](https://github.com/c960657))
 - For Mexico, Constitution Day, Benito Juárez’s birthday, and Revolution Day are considered official holidays. [\#359](https://github.com/azuyalabs/yasumi/pull/359) ([c960657](https://github.com/c960657))
 - For Argentina, Carneval Monday and Tuesday, and Good Friday are considered official holidays. [\#360](https://github.com/azuyalabs/yasumi/pull/360) ([c960657](https://github.com/c960657))
 - For the Czech Republic, Christmas Eve is considered an official holiday. [\#366](https://github.com/azuyalabs/yasumi/pull/366) ([c960657](https://github.com/c960657))